requirements.txt
fastapi
uvicorn
torch==2.2.0+cpu
torchaudio==2.2.0+cpu
transformers==4.38.2
numpy==1.26.4
google-generativeai
--extra-index-url https://download.pytorch.org/whl/cpu
python-multipart

